Fix DSi build
* Fix DSI build. Now works correctly on DSi.
* Fixed soft reset into boot.nds. Now works on both DSi and flashcarts.
Also fixed for theme reset.
* Fixed soft reset into GBA mode. Was using libnds's broken swi call.
Replaced with modified version that isn't broken.
* Fixed warnings generated during compile.
* Removed old LD/Specs from arm7 folder. This is needed to allow proper
arm7 entry point for TWL_FIRM compatiblity on 3DS. Didn't seem to break
DS mode copy so a DSi specific build of arm7 make file is not needed at
the moment.
* custom banner added via banner.bin and -t ndstool command.
* Flashcart build now uses -h 0x200 command for better compatiblity with
older homebrew launchers/flashcarts.

+++ arm9/source/gdi.cpp
@@ -304,7 +304,7 @@ void cGdi::frameRect(s16 x, s16 y, u16 w, u16 h, u16 thickness, GRAPHICS_ENGINE 
 }
 
 void cGdi::fillRect(u16 color1, u16 color2, s16 x, s16 y, u16 w, u16 h, GRAPHICS_ENGINE engine) {
-    ALIGN(4) u16 color[2] = {BIT(15) | color1, BIT(15) | color2};
+    ALIGN(4) u16 color[2] = { (u16)(BIT(15) | color1), (u16)(BIT(15) | color2) };
     u16* pSrc = (u16*)color;
     u16* pDest = NULL;
